Rename common components directory to shared
