+++ app/components/VerseOfDay.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState, useCallback, useRef } from 'react';
 import { Verse, Word } from '@/types';
 import { getRandomVerse } from '@/lib/api';
 import { useSettings } from '@/app/context/SettingsContext';
-import Spinner from '@/app/components/common/Spinner';
+import Spinner from '@/app/components/shared/Spinner';
 import surahsData from '@/data/surahs.json';
 import type { Surah } from '@/types';
 import { useTheme } from '@/app/context/ThemeContext';
