+++ app/features/surah/[surahId]/_components/Verse.tsx
@@ -6,12 +6,12 @@ import {
   FaBookmark,
   FaRegBookmark,
   FaBookReader,
-} from '@/app/components/common/SvgIcons';
+} from '@/app/components/shared/SvgIcons';
 import { useRouter } from 'next/navigation';
 import { Verse as VerseType, Translation, Word } from '@/types';
 import type { LanguageCode } from '@/lib/languageCodes';
 import { useAudio } from '@/app/context/AudioContext';
-import Spinner from '@/app/components/common/Spinner';
+import Spinner from '@/app/components/shared/Spinner';
 import { useSettings } from '@/app/context/SettingsContext';
 import { applyTajweed } from '@/lib/tajweed';
