+++ app/features/tafsir/[surahId]/[ayahId]/_components/VerseCard.tsx
@@ -5,12 +5,12 @@ import {
   FaBookmark,
   FaRegBookmark,
   FaShare,
-} from '@/app/components/common/SvgIcons';
+} from '@/app/components/shared/SvgIcons';
 import { Verse as VerseType, Translation, Word } from '@/types';
 import type { LanguageCode } from '@/lib/languageCodes';
 import { useAudio } from '@/app/context/AudioContext';
 import { useSettings } from '@/app/context/SettingsContext';
-import Spinner from '@/app/components/common/Spinner';
+import Spinner from '@/app/components/shared/Spinner';
 import { applyTajweed } from '@/lib/tajweed';
 import DOMPurify from 'dompurify';
